% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_choropleth.R
\name{data_to_map}
\alias{data_to_map}
\title{Sending map the data.}
\usage{
data_to_map(map, data, map_merge_key = "adm_dr_nm",
  map_key = "adm_dr_cd", data_key = "dong")
}
\arguments{
\item{map}{map}

\item{data}{data mapped}

\item{map_merge_key}{key in map object used to link map object and data.}

\item{map_key}{key used in map object.}

\item{data_key}{kep in data used to link map object and data.}
}
\description{
This function send map object the data which are mapped and change map object to data.frame object.
}
