% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming map.R
\name{name_in_map}
\alias{name_in_map}
\title{naming choropleth.}
\usage{
name_in_map(map, name_var, name = NULL)
}
\arguments{
\item{map}{map object}

\item{name_var}{the variable of name mapped in map object}

\item{name}{name mapped if map object don't have name variable}
}
\description{
This function calculate means latitude and longitude in choropleth.
}
