% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_choropleth.R
\name{plot_map}
\alias{plot_map}
\title{plot the map}
\usage{
plot_map(map, re_var, index_type, path_color = "white", legend_label,
  color_type = "blue_red", color_good = "blue", color_bad = "red",
  fixed_mid = NULL)
}
\arguments{
\item{map}{map having data(data.frame)}

\item{re_var}{variable mapped}

\item{index_type}{1 or 2}

\item{path_color}{color of map boundary}

\item{legend_label}{label of legend}

\item{color_type}{blue_red or viridis}
}
\description{
This function plot choropleth.
}
